+++ main_project/component/currentControlledVoltageSource.hpp
@@ -15,8 +15,8 @@ public:
     CurrentControlledVoltageSource(string _name, vector<string> args, vector<float> extraInfo);
     ~CurrentControlledVoltageSource() = default;
 
-    float ivAtNode(int n) const override;
-    float divAtNode(int n, int dn) const override;
+    double ivAtNode(int n) const override;
+    double divAtNode(int n, int dn) const override;
 
     float getTotalCurrent(const VectorXd &x, int highestNodeNumber, float voltage = 0, int order = 1)  override;
 };
