+++ main_project/component/currentSource.cpp
@@ -63,10 +63,10 @@ vector<int> CurrentSource::getNodes() const{
 }
 
 // iv functions
-float CurrentSource::ivAtNode(int n1) const{
+double CurrentSource::ivAtNode(int n1) const{
 	return current * (n1==nodes[0]?-1:1);
 }
-float CurrentSource::divAtNode(int n1, int dn) const{
+double CurrentSource::divAtNode(int n1, int dn) const{
 	return 0;
 }
