+++ main_project/component/voltageControlledCurrentSource.hpp
@@ -11,8 +11,8 @@ public:
     VoltageControlledCurrentSource(string _name, vector<string> args, vector<float> extraInfo);
     ~VoltageControlledCurrentSource() = default;
 
-    float ivAtNode(int n) const override;
-    float divAtNode(int n, int dn) const override;
+    double ivAtNode(int n) const override;
+    double divAtNode(int n, int dn) const override;
 
     float getTotalCurrent(const VectorXd &x, int highestNodeNumber, float voltage = 0, int order = 1)  override;
 };
