+++ main_project/component/voltageSource.cpp
@@ -62,10 +62,10 @@ vector<int> VoltageSource::getNodes() const{
     return res;
 }
 
-float VoltageSource::ivAtNode(int n) const{
+double VoltageSource::ivAtNode(int n) const{
 	return 1;
 }
-float VoltageSource::divAtNode(int n, int dn) const{
+double VoltageSource::divAtNode(int n, int dn) const{
 	return 1;
 }
